Complete Implementation
