+++ JamGame/Assets/Scenes/Main.unity
@@ -1588,7 +1588,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4727041377538968745, guid: 443513e7dbeafab4ebe5afbc3769efc3, type: 3}
       propertyPath: m_AnchoredPosition.x
-      value: 30
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 4727041377538968745, guid: 443513e7dbeafab4ebe5afbc3769efc3, type: 3}
       propertyPath: m_AnchoredPosition.y
@@ -4559,7 +4559,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 7208807984305182906, guid: d8bbdb0a03f31bd48a52b8d2eb1c238e, type: 3}
       propertyPath: m_AnchoredPosition.x
-      value: 100
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 7208807984305182906, guid: d8bbdb0a03f31bd48a52b8d2eb1c238e, type: 3}
       propertyPath: m_AnchoredPosition.y
