+++ JamGame/Assets/Scripts/Common/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 699af89d6f7cf8147bb07a3c19f3485b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
