+++ JamGame/Assets/Scripts/Common/UI/OnHoverImageColorSwapper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6d7b83d6f3c7ce498f4ce0bf40525db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
