+++ JamGame/Assets/UI/Prefabs/Shop/ShopUI.prefab
@@ -1592,8 +1592,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 1, y: 1}
-  m_AnchoredPosition: {x: 30, y: -35}
-  m_SizeDelta: {x: -60, y: -70}
+  m_AnchoredPosition: {x: 20, y: -35}
+  m_SizeDelta: {x: -40, y: -70}
   m_Pivot: {x: 0, y: 1}
 --- !u!222 &4521669247439445027
 CanvasRenderer:
